requirements.txt
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
black==21.7b0
click==8.0.1
colorama==0.4.4
Django==3.2.6
django-filter==2.4.0
django-mysql==3.12.0
djangorestframework==3.12.4
flake8==3.9.2
iniconfig==1.1.1
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==1.4.2.post1
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.0
